Added health updates for patients

+++ Assets/Scripts/Person.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 
 public class Person : Clickable {
-	protected int health = 100;
+	protected float health = 100;
 
 	// Used when the person has reached an eventual bed
 	protected bool shouldGoToBed = false;
